+++ src/components/ImageLoader/BlurhashCanvas.jsx
@@ -1,12 +1,14 @@
 import { useRef, useState, useEffect } from 'react';
 import { decode } from 'blurhash';
 
-export default ({ style, hash, punch, ratio, width, height }) => {
-  const ref = useRef();
+export default ({ style, hash, punch, ratio, width, height, blurhashRef }) => {
+  if (!blurhashRef) {
+    blurhashRef = useRef();
+  }
   const [styleHeight, setStyleHeight] = useState();
 
   useEffect(() => {
-    const canvas = ref.current;
+    const canvas = blurhashRef.current;
     if (canvas) {
       const pixels = decode(hash, width, height, punch);
       const ctx = canvas.getContext('2d');
@@ -14,25 +16,29 @@ export default ({ style, hash, punch, ratio, width, height }) => {
       imageData.data.set(pixels);
       ctx.putImageData(imageData, 0, 0);
     }
-  }, [hash, width, height, punch]);
+  }, [hash, width, height, punch, blurhashRef]);
 
   useEffect(() => {
-    const canvas = ref.current;
+    const canvas = blurhashRef.current;
     if (canvas) {
-      const adjustHeight = () => setStyleHeight(canvas.offsetWidth / ratio);
-      adjustHeight();
-      const observer = new ResizeObserver(adjustHeight);
-      observer.observe(canvas);
-      return () => observer.unobserve(canvas);
+      if (canvas?.style.aspectRatio) {
+        setStyleHeight('auto');
+      } else {
+        const adjustHeight = () => setStyleHeight(canvas.offsetWidth / ratio);
+        adjustHeight();
+        const observer = new ResizeObserver(adjustHeight);
+        observer.observe(canvas);
+        return () => observer.unobserve(canvas);
+      }
     }
-  }, [ratio]);
+  }, [ratio, blurhashRef, blurhashRef.current?.style.aspectRatio]);
 
   return (
     <canvas
       style={{ ...style, height: styleHeight }}
       height={height}
       width={width}
-      ref={ref}
+      ref={blurhashRef}
     />
   );
 };